requirements.txt
fastapi
uvicorn
gradio
google-genai
python-dotenv
yfinance
torch